requirements.txt
langchain
langchain-core
langchain-openai
openai
langchain-anthropic
langchain-google-genai
google-generativeai
langchain-huggingface
transformers<5.0.0
huggingface-hub
python-dotenv
numpy
scikit-learn
torch
torchvision
langchain-community
ctransformers